Containerizing_MindFuel’s_Quote_Delivery_System/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
greenlet==3.3.0
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.2
